third_party/WebKit/LayoutTests/editing/pasteboard/block-wrappers-necessary-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification


Tests: 

Bug fix for <rdar://problem/4180820> Erroneous spacing between lines introduced on paste

Applying block styles, like text-align, used to wrap paragraphs that needed block style changes in clones of other paragraphs in the document.  This would almost always introduce unwanted styles.


Before cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       "
<#selection-anchor>This is an "
|       <b>
|         "interactive"
|       " documentation site, so please help out by posting any useful information that you've had to search out."
|       <br>
|       "
(Adding to this "
|       <a>
|         href="db.cgi?Wiki_Is_So_Easy"
|         "wiki is so easy"
|       ", you can't screw it up). See "
|       <a>
|         href="db.cgi?Wiki_Reference"
|         "Wiki Reference"
|       " for text formatting syntax."
|       <br>
|       "
Info for setting up your own wiki is at "
|       <a>
|         href="db.cgi?About_Howto.Apple.Com"
|         "about howto.apple.com"
|       "."
|       <br>
|       "
Check out a cool wiki editor application that helps in wiki generation: "
|       <a>
|         href="db.cgi?Wikinator"
|         "Wikinator"
|       "
"
|     <p>
|       "
Start browsing the topics below, or use the search field at the bottom of every page.<#selection-focus>
"
|   "
"
| "
"

After cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       <#selection-caret>
|       <br>
|   "
"
| "
"

After paste (this should be identical to before cut):
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       "This is an "
|       <b>
|         "interactive"
|       " "
|       "documentation site, so please help out by posting any useful information that you've had to search out."
|       <br>
|       "(Adding to this"
|       " "
|       <a>
|         href="db.cgi?Wiki_Is_So_Easy"
|         "wiki is so easy"
|       ", you can't screw it up). See"
|       " "
|       <a>
|         href="db.cgi?Wiki_Reference"
|         "Wiki Reference"
|       " "
|       "for text formatting syntax."
|       <br>
|       "Info for setting up your own wiki is at"
|       " "
|       <a>
|         href="db.cgi?About_Howto.Apple.Com"
|         "about howto.apple.com"
|       "."
|       <br>
|       "Check out a cool wiki editor application that helps in wiki generation:"
|       " "
|       <a>
|         href="db.cgi?Wikinator"
|         "Wikinator"
|     <p>
|       "Start browsing the topics below, or use the search field at the bottom of every page.<#selection-caret>"
|   "
"
| "
"
execCutCommand: <div id="test" class="editing"> <center> <p><br></p></center> </div>
execPasteCommand: <div id="test" class="editing"> <center> <p>This is an&nbsp;<b>interactive</b>&nbsp;documentation site, so please help out by posting any useful information that you've had to search out.<br>(Adding to this&nbsp;<a href="db.cgi?Wiki_Is_So_Easy">wiki is so easy</a>, you can't screw it up). See&nbsp;<a href="db.cgi?Wiki_Reference">Wiki Reference</a>&nbsp;for text formatting syntax.<br>Info for setting up your own wiki is at&nbsp;<a href="db.cgi?About_Howto.Apple.Com">about howto.apple.com</a>.<br>Check out a cool wiki editor application that helps in wiki generation:&nbsp;<a href="db.cgi?Wikinator">Wikinator</a></p><p>Start browsing the topics below, or use the search field at the bottom of every page.</p></center> </div>
